Move stuff to more suited positions.
