+++ pkg/sources/sources.go
@@ -6,29 +6,21 @@
 // SPDX-FileCopyrightText: 2024 German Federal Office for Information Security (BSI) <https://www.bsi.bund.de>
 // Software-Engineering: 2024 Intevation GmbH <https://intevation.de>
 
+// Package sources implements the download from sources.
 package sources
 
 import (
-	"bytes"
 	"context"
-	"crypto/sha256"
-	"crypto/sha512"
-	"encoding/json"
 	"fmt"
-	"hash"
-	"io"
 	"net/http"
 	"net/url"
 	"slices"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/ISDuBA/ISDuBA/pkg/config"
 	"github.com/ISDuBA/ISDuBA/pkg/database"
-	"github.com/ISDuBA/ISDuBA/pkg/models"
 	"github.com/ISDuBA/ISDuBA/pkg/version"
-	"github.com/csaf-poc/csaf_distribution/v3/csaf"
 	"github.com/csaf-poc/csaf_distribution/v3/util"
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
@@ -113,38 +105,6 @@ func (f *feed) refresh(m *Manager) error {
 
 	f.log(m, config.InfoFeedLogLevel, "entries to download: %d", len(f.locations))
 	return nil
-}
-
-// wait establishes the request rate per source.
-func (s *source) wait(ctx context.Context) {
-	if s.rate != nil {
-		s.limiterMu.Lock()
-		defer s.limiterMu.Unlock()
-		if s.limiter == nil {
-			s.limiter = rate.NewLimiter(rate.Limit(*s.rate), 1)
-		}
-		s.limiter.Wait(ctx)
-	}
-}
-
-// doRequest executes an HTTP request with the source specific parameters.
-func (s *source) doRequest(req *http.Request) (*http.Response, error) {
-	req.Header.Add("User-Agent", userAgent)
-
-	client := http.Client{}
-	s.wait(context.Background())
-
-	// TODO: Implement me!
-
-	return client.Do(req)
-}
-
-func (s *source) httpGet(url string) (*http.Response, error) {
-	req, err := http.NewRequest(http.MethodGet, url, nil)
-	if err != nil {
-		return nil, err
-	}
-	return s.doRequest(req)
 }
 
 // fetchIndex fetches the content of the feed index.
@@ -251,130 +211,57 @@ func (f *feed) sameOrNewer() func(*location) bool {
 	}
 }
 
-// download fetches the files of a document and stores
-// them into the database.
-func (l location) download(m *Manager, f *feed, done func()) {
-	defer done()
-
-	var (
-		writers        []io.Writer
-		checksum       hash.Hash
-		remoteChecksum []byte
-	)
-
-	// Loading the hash
-	if l.hash != nil { // ROLIE gave us an URL to hash file.
-		hashFile := l.hash.String()
-		switch lc := strings.ToLower(hashFile); {
-		case strings.HasSuffix(lc, ".sha256"):
-			checksum = sha256.New()
-		case strings.HasSuffix(lc, ".sha512"):
-			checksum = sha512.New()
-		}
-		if checksum != nil {
-			var err error
-			if remoteChecksum, err = f.source.loadHash(hashFile); err != nil {
-				f.log(m, config.WarnFeedLogLevel, "fetching hash %q failed: %v", hashFile, err)
-			} else {
-				writers = append(writers, checksum)
-			}
-		}
-	} else if !f.rolie { // If we are directory based, do some guessing:
-		for _, h := range []struct {
-			ext  string
-			cstr func() hash.Hash
-		}{
-			{".sha512", sha512.New},
-			{".sha256", sha256.New},
-		} {
-			guess := l.doc.String() + h.ext
-			if rc, err := f.source.loadHash(guess); err == nil {
-				remoteChecksum, checksum = rc, h.cstr()
-				writers = append(writers, checksum)
-				break
-			}
+// findLocationByID looks for location with a given id.
+func (f *feed) findLocationByID(id int64) *location {
+	for i := len(f.locations) - 1; i >= 0; i-- {
+		if location := &f.locations[i]; location.id == id {
+			return location
 		}
 	}
+	return nil
+}
 
-	// Download the CSAF document.
-	resp, err := f.source.httpGet(l.doc.String())
+// findWaiting looks for a location ready to download.
+func (f *feed) findWaiting() *location {
+	// Backwards because the new ones are at the end.
+	for i := len(f.locations) - 1; i >= 0; i-- {
+		if location := &f.locations[i]; location.state == waiting {
+			return location
+		}
+	}
+	return nil
+}
+
+// wait establishes the request rate per source.
+func (s *source) wait(ctx context.Context) {
+	if s.rate != nil {
+		s.limiterMu.Lock()
+		defer s.limiterMu.Unlock()
+		if s.limiter == nil {
+			s.limiter = rate.NewLimiter(rate.Limit(*s.rate), 1)
+		}
+		s.limiter.Wait(ctx)
+	}
+}
+
+// doRequest executes an HTTP request with the source specific parameters.
+func (s *source) doRequest(req *http.Request) (*http.Response, error) {
+	req.Header.Add("User-Agent", userAgent)
+
+	client := http.Client{}
+	s.wait(context.Background())
+
+	// TODO: Implement me!
+
+	return client.Do(req)
+}
+
+func (s *source) httpGet(url string) (*http.Response, error) {
+	req, err := http.NewRequest(http.MethodGet, url, nil)
 	if err != nil {
-		f.log(m, config.ErrorFeedLogLevel, "downloading %q failed: %v", l.doc, err)
-		return
+		return nil, err
 	}
-	if resp.StatusCode != http.StatusOK {
-		f.log(m, config.ErrorFeedLogLevel, "downloading %q failed: %s (%d)",
-			l.doc, http.StatusText(resp.StatusCode), resp.StatusCode)
-		return
-	}
-	defer resp.Body.Close()
-
-	var data bytes.Buffer
-	writers = append(writers, &data)
-
-	// Prevent over-sized downloads.
-	limited := io.LimitReader(resp.Body, int64(m.cfg.General.AdvisoryUploadLimit))
-
-	tee := io.TeeReader(limited, io.MultiWriter(writers...))
-
-	// Decode document into JSON.
-	var doc any
-	if err := json.NewDecoder(tee).Decode(&doc); err != nil {
-		f.log(m, config.ErrorFeedLogLevel, "decoding document %q failed: %v", l.doc, err)
-		return
-	}
-
-	// Compare checksums.
-	if remoteChecksum != nil {
-		if !bytes.Equal(checksum.Sum(nil), remoteChecksum) {
-			f.log(m, config.ErrorFeedLogLevel, "Checksum mismatch for document %q", l.doc)
-			return
-		}
-	}
-
-	// Check document against schema.
-	if errors, err := csaf.ValidateCSAF(doc); err != nil || len(errors) > 0 {
-		if err != nil {
-			f.log(m, config.ErrorFeedLogLevel,
-				"Schema validation of document %q failed: %v", l.doc, err)
-		} else {
-			f.log(m, config.ErrorFeedLogLevel,
-				"Schema validation of document %q has %d errors", l.doc, len(errors))
-		}
-		return
-	}
-
-	// TODO: Check against remote validator
-	// TODO: Filename check. (???)
-	// TODO: Check signatures
-	// TODO: Statistics
-
-	// Remember last changes.
-	inTx := func(ctx context.Context, tx pgx.Tx, _ int64) error {
-		const updatedSQL = `INSERT INTO changes (url, feeds_id, time) ` +
-			`VALUES ($1, $2, $3) ` +
-			`ON CONFLICT (url, feeds_id) DO ` +
-			`UPDATE SET time = $3`
-		_, err := tx.Exec(ctx, updatedSQL, l.doc.String(), f.id, l.updated)
-		return err
-	}
-
-	ctx := context.Background()
-	if err := m.db.Run(ctx, func(ctx context.Context, conn *pgxpool.Conn) error {
-		_, err := models.ImportDocumentData(
-			ctx, conn,
-			doc, data.Bytes(),
-			m.cfg.Sources.FeedImporter,
-			m.cfg.Sources.PublishersTLPs,
-			inTx,
-			false)
-		return err
-	}, 0); err != nil {
-		f.log(m, config.ErrorFeedLogLevel, "storing %q failed: %v", l.doc, err)
-		return
-	}
-
-	f.log(m, config.InfoFeedLogLevel, "downloading %q done", l.doc)
+	return s.doRequest(req)
 }
 
 // loadHash fetches text form of a hash from remote location.
@@ -387,4 +274,18 @@ func (s *source) loadHash(url string) ([]byte, error) {
 		return nil, fmt.Errorf("%s (%d)", http.StatusText(resp.StatusCode), resp.StatusCode)
 	}
 	return util.HashFromReader(resp.Body)
+}
+
+// storeLastChanges is intented to be called in the transaction storing the
+// importing the document after is was successful. It helps to remember the
+// last changes per location so we don't need to download them all again and again.
+func (f *feed) storeLastChanges(l *location) func(context.Context, pgx.Tx, int64) error {
+	return func(ctx context.Context, tx pgx.Tx, _ int64) error {
+		const updatedSQL = `INSERT INTO changes (url, feeds_id, time) ` +
+			`VALUES ($1, $2, $3) ` +
+			`ON CONFLICT (url, feeds_id) DO ` +
+			`UPDATE SET time = $3`
+		_, err := tx.Exec(ctx, updatedSQL, l.doc.String(), f.id, l.updated)
+		return err
+	}
 }
